move: view_filter_offer_full
