+++ app/routers/consts.py
@@ -5,3 +5,4 @@ from dataclasses import dataclass
 class RouteSlug:
     pk: str = '{pk}/'
     ifilter: str = '{ifilter}/'
+    full: str = 'full/'
